pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["Behrooz Payoon <Behrooz.payoon@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
alembic = "1.7.7"
passlib = {version = "1.7.4", extras = ["bcrypt"]}
pydantic = {version = "1.9.1", extras = ["email"]}
fastapi = "0.79.0"
uvicorn = "0.18.2"
gunicorn = "20.1.0"
psycopg2 = "2.9.3"
tenacity = "8.0.1"
PyJWT = "2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
